Broke up MatchList into multiple files since it handled a lot of different things, and was becoming quiet long and complex
